Add files via upload
Here we have done our research on moss ceratodon purpures which is a bryophyte. It was found in devarayanadurga hills which was then taken to the laboratory and tested for its RWC(relative water content) in hydrated(H) dehydrated(D) and rehydrated(R) state and tested its chlorophyll content , proline content etc by using byochemical assays. The results and the procedure used for the experiment is written in detail in the above research paper.
The writing rights are co-authored by Vidya M, Yashaswini P, Rakshitha K Nayak.